Improved controller support
Standard layout game pads should now be supported alongside extended layout gamepads.

The SNES layout will put 'start' and 'select' on the triggers of the extended gamepad, on the standard layout they'll be left on-screen.

For the Genesis layout, the extra 3 buttons for the 6-button controller will be mapped as:

left-shoulder: X
Y: Y
right-shoulder: Z

A, B and C are mapped as normal to X, A and B.

(Apple's gamepad layouts are so unconventional).