Much improved game lookup
Provenance will first try to look a ROM up by its MD5 hash. If that fails it will try the ROM's CRC32 hash. If that fails too, then it will fall back to a search based on the ROM's file name. If there is a USA entry it will choose that one, if not, it will choose the first entry in the results. Of course there will always be exceptions and some ROMs may never get found.
The name-based search also factors in the system type, so it should be pretty accurate.
This solves issue #62 (NES Artworks), and generally improves artwork look up for many ROMs that don't immediately match by their hash.
I added a button in the settings to refresh your game library, so that you can try and get art for ROMs you previously had nothing for.