Copied over the dpad improvements made by kckd to all cores. Seems to improve dpad handling across the board.
Also changed Siri Remote mappings. Touchpad click is equivalent to "left trigger" on a game pad, which is usually mapped to Start on most systems. The play/pause button is now mapped to A. Sonic and Mario are now kind of playable with the Siri Remote, that's all that matters right?!