Restore `frameRefreshThread:`'s implementation to the base class
Huge thanks to *astrange* from OpenEmu for pointing out that `init_mach_time()` wasn't wrapped in a `dispatch_once` call, which was leading to race conditions. I still don't fully understand why it works in the subclass and not the base class without the dispatch_once, but I'm happy that the duplicated code is now gone.