Fix #311... again.
stupid bug in tvOS 9.2

the controller paused handler (if implemented) seems to cause a 'back' navigation action as well as calling the pause handler itself. Which breaks the menu functionality.

But of course, this isn't the case on iOS 9.3. YAY FRAGMENTATION. ¬_¬

Conditionally handle the pause menu differently dependning on tvOS or iOS. FFS.