Show current and master SHA hashes in Settings
Adds new Build Phases to create local.plist (which is ignored by git)
containing the SHA of the current branch's most recent commit and the SHA
of the commit in master from which the local branch was created. The latter
updates after rebasing against an updated master branch.

Both of these SHAs can then be seen on the Settings menu, which may be
helpful for both developers and users (who may be having issues building and
installing Provenance.