Fix audio crash on second load
The 800 core initilaizes the var Sound_enabled to true on it’s static, but then sets to false on deinit. On the next init, unless you set ‘-sound’ argv that static is still false on reinit. Which cuaes it to not initialize it’s sound buffers. When we go to reneder sound it crashes on the second time the core is used. Why they init a static to 1 and later reset it to 0 I have no fucking clue.