reworked the screenshot capturing
iPhones would crash because the GL view was removed from the heirarchy when the save states view controller was presented, meaning a screenshot was unable to be rendered. Now we take a screenshot before showing the save states UI, and use that screenshot for any states created manually then. Auto save states, continue to take their screenshot when they’re created.