Fixes crash that can happen when launching a game.
The Recently Saved section only displays the last 6 items, but the user probably has way more states saved than that. So when the collectionView sections update, sometimes the Realm update notification wants to update items outside the first 6, but those items don't exist in the collectionView. This causes a crash.

To fix this, we filter the deletions, insertions, and modifications to only those within the first 6 items.