Fix issues with ZipArchive.framework breaking ad-hoc signing
It should have been linked o PVLibrary, but it was only being linked in a emved stage in the app target. It’s a carthage framework so it needs to be in the scripted stage to strip it of extra stuff that breaks store/adhoc exports.

Also shouldn’t have been in the swift bridge header, it can be natively imported in either language as a module import.