First pass at working ish VGC
need to re-add listeners for iCade. Text view doesn’t work at all so no iCade inputs.

Need to add a way to ask the user what kind of iCade it is.
Also need to support more than 1 iCade somehow in the future (just add 8bitdo player 2 as a type?).

Need to make a better interface so that the cores don’t need to know about GC or VGC classes