Don't call view.isHidden on unloaded view, fixes #1066.
I'm not really sure why this call is here, but removing it seems to fix
the rather strange issue #1066. Anyway, since the codePath where
`saveState == nil` doesn't ever run `view.isHidden = false` it seems
like this is an error, and I couldn't see any reason why it should be
there in the first place.