Fix #1436 crash on iPad
This is a funny tidbit when using UIImagePicker on iPads (and how action sheets are being presented on iPad). Unlike on iPhones, on iPads it is being presented within a popover and popovers need certain information about the environment they're being presented in (where should the arrow point to, which view is the source of the interaction...).