saveGame titleLabels now match other Labels on tvOS
Sets default state to match the darkGray textColor of the other labels.
Animate on focus turns the labels white for optimal readability like the other labels.
When shifting focus to another title the color returns to the default darkGray.
Match scale and offset when focussing to the behaviour the other Titles have.