Add 8Bitdo M30 Controller support for Genesis System (#1500)
* Enables 8BitDo M30 Mapping for Sega games

This remaps the MFI Extended controllers to better fit to a 8BitDo M30 for use with the Genesis/Megadrive/Sega/MegaCD system inside of Provenance.

All of the buttons on the M30 map 1:1 to where they would be on the 6 Button Sega Controller, it also moves the Start button from the Extended Controller to the Start button on the M30. The L and R triggers are NOT used and enabled.

* Fix Start Button on M30 for iOS

The M30 "Start" button on iOS calls up the in-game Provenance Menu, on tvOS it does not and works as expected. Add in quick #if to catch this and use the R Shoulder button as the "Start" button on iOS.

* Add "mode" button support ....

to Sega Genesis/MD/SegaCD/MegaCD control mapping.