m68k.cpp compile time tweak
This is a quick and non code invasive Mednafen Project compilation tweak to minimize the impact of the m68k.cpp cpu module on the Project compilation time. 
NONE of the cores/modules we use in Mednafen appear to use the m68k cpu anyways.

With this simple file compilation tweak in place on ay MacBookPro 2019 Intel the total compilation time for a vanilla iOS Provenance Release build came down to 326.7seconds ( 5mins 27secs ) from our current vanilla build time of 616.9seconds  ( 10mins and 16secs).