Update PVGameMoreInfoViewController.swift
tvOS Only - This allows the user to use a double click on the controller/remote to view the Back of the Box artwork when in the GameInfo view. 
This also removes the scrolling option and sets the Game Description Text to a fixed 24 System Font for better readability without having to scroll.