Sets require restart to false for a few core options for emuThree; adds description for Shader Accurate Mul
I noticed that most options that have 'requiresRestart: true' do not actually apply on restart. These changes should (in theory) fix upscaling not updating, cpu clock not updating, as well as some other core options. On Citra, most settings do not require a restart anyway, except for JIT.

This needs to be tested as I cannot build emuThree for myself.